Move TestFrontJsonFapi into common/test
We want to be able to call it both from `facia/test` (where it is already used) and `article/test`. This will allow us to write unit tests for `article/app/agents/CuratedContentAgent`.

This PR also moves `TestFrontJsonFapi` to a new `WithTestFrontJsonFapi` trait so that we avoid repetition when using it.
