+++ facia/test/FaciaMetaDataTest.scala
@@ -26,11 +26,8 @@ import test._
     with WithTestApplicationContext
     with WithMaterializer
     with WithTestWsClient
-    with MockitoSugar {
-
-  lazy val actorSystem = ActorSystem()
-  lazy val blockingOperations = new BlockingOperations(actorSystem)
-  lazy val fapi = new TestFrontJsonFapi(blockingOperations)
+    with MockitoSugar
+    with WithTestFrontJsonFapi {
 
   override def beforeAll(): Unit = {
     val refresh = ConfigAgent.refreshWith(
