+++ facia/test/package.scala
@@ -21,34 +21,6 @@ object `package` {
     def hasAttribute(name: String): Boolean = element.attribute(name) != null
   }
 
-  // need a front api that stores S3 locally so it can run without deps in the unit tests
-  class TestFrontJsonFapi(override val blockingOperations: BlockingOperations)
-      extends FrontJsonFapiLive(blockingOperations) {
-
-    override def get(path: String, pageType: PressedPageType)(implicit
-        executionContext: ExecutionContext,
-    ): Future[Option[PressedPage]] = {
-      recorder.load(path, Map()) {
-        super.get(path, pageType)
-      }
-    }
-
-    val recorder = new HttpRecorder[Option[PressedPage]] {
-      override lazy val baseDir = new File(System.getProperty("user.dir"), "data/pressedPage")
-
-      //No transformation for now as we only store content that's there.
-      override def toResponse(b: Array[Byte]): Option[PressedPage] =
-        Json.parse(new String(b, UTF8.charSet)).asOpt[PressedPage]
-
-      override def fromResponse(maybeResponse: Option[PressedPage]): Array[Byte] = {
-        val response = maybeResponse getOrElse {
-          throw new RuntimeException("seeing None.get locally? make sure you have S3 credentials")
-        }
-        Json.stringify(Json.toJson(response)).getBytes(UTF8.charSet)
-      }
-    }
-  }
-
 }
 
 class FaciaTestSuite
